extract useTopStories hook
