+++ www/src/pages/TopStories/PageTopStories.tsx
@@ -1,44 +1,25 @@
 import { StoryCard } from "@/components";
-import { useServices } from "@/contexts";
-import { useEffect, useState } from "react";
-import { StoryCardContainer } from "./components";
-import { Stack } from "@mui/material";
 import { range } from "@/lib/range";
 import styled from "@emotion/styled";
+import { Stack } from "@mui/material";
+import { StoryCardContainer } from "./components";
+import { useTopStories } from "./hooks";
 
 const PageContent = styled.main`
     display: flex;
     justify-content: center;
     width: 100%;
+    height: 100vh;
     padding-top: 2rem;
 `;
 
+export const PAGE_TOP_STORIES_ID = "page-top-stories";
+
 export const PageTopStories = () => {
-    const [isLoading, setLoading] = useState(false);
-    const [topStoriesIds, setTopStoriesIds] = useState<number[]>([]);
-    const { apiAdapter } = useServices();
-
-    const getData = async () => {
-        setLoading(true);
-
-        const topStoriesData = await apiAdapter.getTopStories();
-
-        // show top 200 stories
-        setTopStoriesIds(topStoriesData.slice(0, 200));
-
-        setLoading(false);
-    };
-
-    useEffect(() => {
-        getData();
-
-        return () => {
-            setTopStoriesIds([]);
-        };
-    }, []);
+    const { isLoading, topStoriesIds } = useTopStories();
 
     return (
-        <PageContent>
+        <PageContent id={PAGE_TOP_STORIES_ID}>
             <Stack spacing={2}>
                 {isLoading ? (
                     <>
